Add asynchronous replication (#12)
Adds configurable replication, if there are any replicas configured, the
the primary node will start to append all write operations(SET, INCR and
DECR) to an operation log, this log is read by a background process
to send all operations to the configured replicas. This process can run
concurrently with other write operations because the log is only cleared
after all operations are sent to the replicas. If there's an error
sending the operation to any replica it'll be discarted and dropped from
the log, there are no guarantees that the whole log will be sent to the
replica.